requirements.txt
# Django + DB
Django>=5.0
psycopg[binary]>=3.1
dj-database-url>=2.2
python-dotenv>=1.0
gunicorn>=23.0

# Dev / test / tooling (optional in prod)
black>=24.0
ruff>=0.8
pytest-django>=4.8
pre-commit>=3.0
